% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{activate}
\alias{activate}
\alias{active}
\alias{active<-}
\alias{activate}
\alias{activate.NetCDF}
\alias{activate}
\alias{activate.hyperfilter}
\alias{activate}
\alias{active.NetCDF}
\alias{activate}
\alias{active.hyperfilter}
\alias{activate}
\alias{active<-.NetCDF}
\alias{activate}
\alias{active<-.hyperfilter}
\title{#' @name hyper_filter
#' @export
hyper_filter_Spatial <- function(x, y, ...) {
  warning("assuming first two dimensions are longitude-latitude ...")
  dim_tabs <- hyper_filter(x)
  xy_names <- names(dim_tabs)[1:2]
  xy_grid <- as.matrix(expand.grid(x = dim_tabs[[1]][[xy_names[1]]], 
                         y = dim_tabs[[1]][[xy_names[1]]]))
  over(y, SpatialPoints(xy_grid, proj4string = crs(y)))
}
Activate}
\usage{
activate(.data, what)

\method{activate}{NetCDF}(.data, what)

\method{activate}{hyperfilter}(.data, what)

\method{active}{NetCDF}(x)

\method{active}{hyperfilter}(x)

\method{active}{NetCDF}(x) <- value

\method{active}{hyperfilter}(x) <- value
}
\arguments{
\item{.data}{NetCDF object}

\item{what}{name of a variable}

\item{x}{NetCDF object}

\item{value}{name of variable to be active}

\item{value}{name of variable to be active}
}
\value{
NetCDF object
}
\description{
Set the context for subsequent manipulations.
}
\details{
`activate` puts the named variable first
`active` gets and sets the active variable
}
\examples{
rnc <- ncdump::NetCDF(system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package= "ncdump"))
activate(rnc, "palette")
}
