% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{as.tbl_cube}
\alias{as.tbl_cube}
\alias{as.tbl_cube.tidync_array}
\alias{as.tbl_cube}
\alias{as.tbl_cube.tidync}
\title{Dplyr 'tbl' cubes}
\usage{
\method{as.tbl_cube}{tidync_array}(x, ...)

\method{as.tbl_cube}{tidync}(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{Arguments passed on to \code{dplyr::as.tbl_cube}
\describe{
  \item{x}{an object to convert. Built in methods will convert arrays,
tables and data frames.}
}}
}
\description{
Convert a tidync object to a dplyr tbl_cube
}
