% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{as_tibble.tidync}
\alias{as_tibble.tidync}
\title{Tibble}
\usage{
\method{as_tibble}{tidync}(x, ..., na.rm = TRUE)
}
\arguments{
\item{x}{dplyr object}

\item{...}{}
}
\description{
Convert a tidync object to a tibble (data frame).
}
