% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{axis_transforms}
\alias{axis_transforms}
\alias{axis_transforms}
\alias{axis_transforms.tidync}
\title{Axis transforms}
\usage{
axis_transforms(x, ...)

\method{axis_transforms}{tidync}(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{ignored}
}
\value{
list of axis transforms
}
\description{
Axis transforms
}
