% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{deprecated-tidync}
\alias{deprecated-tidync}
\alias{axis_transforms}
\title{Deprecated functions in tidync}
\usage{
axis_transforms(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{ignored}
}
\description{
Deprecated functions in tidync
}
