% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{dplyr-tidync}
\alias{dplyr-tidync}
\alias{filter.tidync}
\alias{dplyr-tidync}
\alias{collect.tidync}
\title{dplyr methods}
\usage{
\method{filter}{tidync}(.data, ...)

\method{collect}{tidync}(x, ...)
}
\arguments{
\item{.data}{tidync object}

\item{...}{arguments passed to \code{hyper_filter} (for filter) or \code{hyper_array} (for collect)}

\item{x}{tidync object}
}
\description{
dplyr methods
}
