% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_compound.R
\name{h5_vars}
\alias{h5_vars}
\alias{h5_vars}
\alias{h5_vars.character}
\alias{h5_vars}
\alias{h5_vars.H5IdComponent}
\title{Title}
\usage{
h5_vars(x)

\method{h5_vars}{character}(x)

\method{h5_vars}{H5IdComponent}(x)
}
\arguments{
\item{x}{}
}
\description{
Title
}
\examples{
f <- system.file( "extdata", "h5", "S2008001.L3b_DAY_RRS.nc", package = "tidync")
nc_groups(f)
a <- h5_vars(f)
handle <- rhdf5::H5Fopen(f)
x <- rhdf5::h5read(handle, sprintf("/\%s", a$name[1]))
names(x)
## roc only cares about 1:10 here, the rest are guff
## those 10 names are data frames, and BinIndex is common to NROWs (e.g. MODISA)
## and the rest are the same size, BinList in the bin_num and weights etc
## and the Rrs/angstromg/aot etc are the sums and ssq - so really it's only one data frame
## from the hypertidy perspective
[1] "BinIndex"     "BinList"      "Rrs_412"      "Rrs_443"      "Rrs_490"      "Rrs_510"     
[7] "Rrs_555"      "Rrs_670"      "angstrom"     "aot_865"      "binDataDim"   "binDataType" 
[13] "binIndexDim"  "binIndexType" "binListDim"   "binListType" 
}
