% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-groupby.R
\name{hyper_group_by}
\alias{hyper_group_by}
\title{Hyper group by}
\usage{
hyper_group_by(x, object, ns = 1:2)
}
\arguments{
\item{x}{a tidync object}

\item{object}{a polygon object}

\item{ns}{the dimensions of the active grid to use as the nominal space}
}
\description{
Group the nominated space within a grid by an polygon object
}
