% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{hyper_index}
\alias{hyper_index}
\alias{hyper_index}
\alias{hyper_index.character}
\alias{hyper_index}
\alias{hyper_index.tidync}
\title{hyper index}
\usage{
hyper_index(x, ..., select_var = NULL, raw_datavals = FALSE)

\method{hyper_index}{character}(x, ..., select_var = NULL,
  raw_datavals = FALSE)

\method{hyper_index}{tidync}(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{ignored}

\item{select_var}{optional vector of variable names to select}

\item{raw_datavals}{logical to control whether scaling in the NetCDF is applied or not}
}
\value{
tidync object
}
\description{
Perform a traditional unnamed array slice operation. This assumes that
no previous selection has been made. This is a lazy verb, we still need
hyper_array or collect to actuall pull the data.
}
