% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{hyper_slice}
\alias{hyper_slice}
\alias{hyper_slice}
\alias{hyper_slice.hyperindex}
\alias{hyper_slice}
\alias{hyper_slice.hyperfilter}
\alias{hyper_slice}
\alias{hyper_slice.tidync}
\alias{hyper_slice}
\alias{hyper_slice.character}
\title{hyper slice}
\usage{
hyper_slice(x, ..., raw_datavals = FALSE)

\method{hyper_slice}{hyperindex}(x, ..., raw_datavals = FALSE)

\method{hyper_slice}{hyperfilter}(x, ..., raw_datavals = FALSE)

\method{hyper_slice}{tidync}(x, ..., raw_datavals = FALSE)

\method{hyper_slice}{character}(x, ..., raw_datavals = FALSE)
}
\arguments{
\item{x}{a hyperindex-able thing}

\item{...}{ignored}

\item{raw_datavals}{logical to control whether scaling in the NetCDF is applied or not}
}
\description{
hyper slice
}
