% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr-pipe.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\examples{
system.file("extdata/argo/MD5903593_001.nc", package = "tidync") \%>\% 
     tidync()
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

