% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_tidyfile.R
\name{tidyfile}
\alias{tidyfile}
\alias{tidyfile}
\alias{tidyfile.data.frame}
\alias{tidyfile}
\alias{tidyfile.raadfiles}
\alias{tidyfile}
\alias{tidyfile.character}
\title{Tidy file collection}
\usage{
tidyfile(x, all = TRUE, ...)

\method{tidyfile}{data.frame}(x, all = TRUE, ...)

\method{tidyfile}{raadfiles}(x, all = TRUE, ...)

\method{tidyfile}{character}(x, all = TRUE, ...)
}
\arguments{
\item{x}{raadfiles, data frame with "fullname", or character vector}

\item{all}{scan metadata from all files?}

\item{...}{ignored}
}
\value{
data frame of filename "fullname" and "meta" in a list-col
}
\description{
Tidy file collection
}
\examples{
#library(dplyr)
#files <- raadfiles::oisst_daily_files() \%>\% dplyr::filter(as.Date(date) > (Sys.Date() - 40))
#f <- tidyfile(files, all = FALSE)
}
