% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{variable_dimensions}
\alias{variable_dimensions}
\alias{variable_dimensions}
\alias{variable_dimensions}
\title{Dimensions of a variable}
\usage{
variable_dimensions(x)

variable_dimensions(x)

variable_dimensions(x)
}
\arguments{
\item{x}{NetCDF object}
}
\description{
Dimensions of a variable
}
