% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\alias{active}
\alias{active<-}
\alias{activate}
\alias{activate.tidync}
\alias{activate}
\alias{activate.hyperfilter}
\alias{activate}
\alias{active.tidync}
\alias{activate}
\alias{active.hyperfilter}
\alias{activate}
\alias{active<-.tidync}
\alias{activate}
\alias{active<-.hyperfilter}
\title{Activate}
\usage{
activate(.data, what)

\method{activate}{tidync}(.data, what)

\method{activate}{hyperfilter}(.data, what)

\method{active}{tidync}(x)

\method{active}{hyperfilter}(x)

\method{active}{tidync}(x) <- value

\method{active}{hyperfilter}(x) <- value
}
\arguments{
\item{.data}{NetCDF object}

\item{what}{name of a variable}

\item{x}{NetCDF object}

\item{value}{name of variable to be active}
}
\value{
NetCDF object
}
\description{
Set the context for subsequent manipulations.
}
\details{
\code{activate} puts the named variable first
\code{active} gets and sets the active variable
}
\examples{
l3file <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package= "ncdump")
rnc <- tidync(l3file)
activate(rnc, "palette")
}
