% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\alias{active}
\alias{active<-}
\alias{activate}
\alias{activate.tidync}
\alias{activate}
\alias{active.tidync}
\alias{active}
\alias{active.default}
\alias{active<-}
\alias{active<-.default}
\title{Activate}
\usage{
activate(.data, what, ..., select_var = NULL)

\method{activate}{tidync}(.data, what, ..., select_var = NULL)

\method{active}{tidync}(x)

active(x)

\method{active}{default}(x)

active(x) <- value

\method{active}{default}(x) <- value
}
\arguments{
\item{.data}{NetCDF object}

\item{what}{name of a variable}

\item{x}{NetCDF object}

\item{value}{name of grid or variable to be active}
}
\value{
NetCDF object
}
\description{
Set the context for subsequent manipulations.
}
\details{
\code{activate} puts the named variable first
\code{active} gets and sets the active variable
}
\examples{
l3file <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
rnc <- tidync(system.file("extdata", "oceandata", l3file, package = "tidync"))
activate(rnc, "palette")
}
