% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{axis_transforms}
\alias{axis_transforms}
\alias{axis_transforms}
\alias{active_axis_transforms}
\alias{axis_transforms}
\alias{axis_transforms.default}
\title{Axis transforms}
\usage{
axis_transforms(x, ...)

active_axis_transforms(x, ...)

\method{axis_transforms}{default}(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{ignored}

\item{active_only}{return all axes (default) or only the ones in the active grid}
}
\value{
list of axis transforms
}
\description{
Axis transforms
}
