% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{axis_transforms}
\alias{axis_transforms}
\alias{axis_transforms}
\alias{active_axis_transforms}
\alias{axis_transforms}
\alias{axis_transforms.default}
\title{Axis transforms}
\usage{
axis_transforms(x, ...)

active_axis_transforms(x, ...)

\method{axis_transforms}{default}(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{ignored}
}
\value{
list of axis transforms
}
\description{
Axis 'transforms' are data frames of each dimension in a NetCDF source.
}
\details{
Each transform is available by name from a list, and each data frame has
the coordinate of the dimension, its index, and a 'selected' variable
set by the filtering expressions in \code{hyper_filter} and used by the
read-functions \code{hyper_array} and \code{hyper_tibble}.

Use \code{axis_transforms} to interrogate and explore the available dimension manually, or
for development of custom functions.
}
\examples{
l3file <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
f <- system.file("extdata", "oceandata", l3file, package = "tidync")
ax <- tidync(f) \%>\% axis_transforms()
names(ax)
lapply(ax, dim)

## this function returns the transforms tidync knows about for this source
str(tidync(f)$transforms)
}
