% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{dim_names}
\alias{dim_names}
\alias{dim_names}
\alias{dim_names.character}
\alias{dim_names}
\alias{dim_names.tidync}
\title{Dimension names}
\usage{
dim_names(x, ...)

\method{dim_names}{character}(x, ...)

\method{dim_names}{tidync}(x, ...)
}
\arguments{
\item{x}{NetCDF object}

\item{...}{ignored}
}
\value{
names of available dimensions
}
\description{
Dimension names
}
\examples{
rnc <- ncdump::NetCDF(system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package= "ncdump"))
dim_names(rnc)
}
