% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{dimension_values}
\alias{dimension_values}
\alias{dimension_values.character}
\alias{dimension_values.tidync}
\title{Dimension values}
\usage{
dimension_values(x)

\method{dimension_values}{character}(x)

\method{dimension_values}{tidync}(x)
}
\arguments{
\item{x}{NetCDF object}
}
\value{
data frame of dimensions and their values
}
\description{
The dimension values are the coordinates, the positions along each axis.
}
\examples{
l3file <- "S2008001.L3m_DAY_CHL_chlor_a_9km.nc"
rnc <- tidync(system.file("extdata", l3file, package= "ncdump"))
dimension_values(rnc)
}
