% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{hyper_array}
\alias{hyper_array}
\title{Hyper slice}
\usage{
hyper_array(x, select_var = NULL, ..., raw_datavals = FALSE,
  force = FALSE)
}
\arguments{
\item{x}{tidync object}

\item{select_var}{optional vector of variable names to select}

\item{...}{ignored}

\item{raw_datavals}{logical to control whether scaling in the NetCDF is applied or not}

\item{force}{ignore caveats about large extraction and just do it}
}
\description{
Slice out a "hyper" array, returning a list of variables in native
array form.
}
