% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_filter.R
\name{hyper_filter}
\alias{hyper_filter}
\alias{hyper_filter}
\alias{hyper_filter.tidync}
\title{Array subset by expression}
\usage{
hyper_filter(.x, ...)

\method{hyper_filter}{tidync}(.x, ...)
}
\arguments{
\item{.x}{NetCDF file, connection object, or \code{tidync} object}

\item{...}{currently ignored}
}
\value{
data frame
}
\description{
Filter arguments must be named as per the dimensions in the variable. This is a restrictive variant of \code{dplyr::filter},
with a syntax more like \code{dplyr::mutate}. This ensures that each element is named, so we know which dimension to
apply this to, but also that the expression evaluated against can do some extra work for a nuanced test.
}
\details{
There are special columns provided with each axis, one is 'index' so that exact matching can be
done by position, or to ignore the actual value of the coordinate.
}
\examples{
f <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
l3file <- system.file("extdata/oceandata", f, package= "tidync")
## filter by value
tidync(l3file) \%>\% hyper_filter(lon = lon < 100)
## filter by index
tidync(l3file) \%>\% hyper_filter(lon = index < 100)

## filter in combination/s
tidync(l3file) \%>\% hyper_filter(lat = abs(lat) < 10, lon = index < 100)
}
