% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{hyper_filter}
\alias{hyper_filter}
\alias{hyper_filter}
\alias{hyper_filter.tidync}
\alias{hyper_filter}
\alias{hyper_filter.default}
\alias{hyper_filter}
\alias{hyper_filter.character}
\alias{hyper_filter}
\alias{hyper_filter.hyperfilter}
\alias{hyper_filter}
\alias{print.hyperfilter}
\title{Array subset by nse}
\usage{
hyper_filter(x, ...)

\method{hyper_filter}{tidync}(x, ...)

\method{hyper_filter}{default}(x, ...)

\method{hyper_filter}{character}(x, ...)

\method{hyper_filter}{hyperfilter}(x, ...)

\method{print}{hyperfilter}(x, ...)
}
\arguments{
\item{x}{NetCDF object}

\item{...}{currently ignored}
}
\value{
data frame
}
\description{
NSE arguments must be named as per the dimensions in the variable. This is a restrictive variant of `dplyr::filter`, 
with a syntax more like `dplyr::mutate`. This ensures that each element is named, so we know which dimension to 
apply this to, but also that the expression evaluated against can do some extra work for a nuanced test.
}
\examples{
## inst/dev/filtrate_early.R
##http://rpubs.com/cyclemumner/tidyslab
# 
}
