% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_filter.R
\name{hyper_filter}
\alias{hyper_filter}
\alias{hyper_filter}
\alias{hyper_filter.tidync}
\title{Array subset by expression}
\usage{
hyper_filter(.x, ...)

\method{hyper_filter}{tidync}(.x, ...)
}
\arguments{
\item{.x}{NetCDF file, connection object, or \code{tidync} object}

\item{...}{currently ignored}
}
\value{
data frame
}
\description{
Filter arguments must be named as per the dimensions in the variable. This is a restrictive variant of \code{dplyr::filter},
with a syntax more like \code{dplyr::mutate}. This ensures that each element is named, so we know which dimension to
apply this to, but also that the expression evaluated against can do some extra work for a nuanced test.
}
\details{
There are special columns provided with each axis, one is 'index' so that exact matching can be
done by position, or to ignore the actual value of the coordinate.
}
\examples{
## inst/dev/filtrate_early.R
##http://rpubs.com/cyclemumner/tidyslab
# 
}
