% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_index.R
\name{hyper_index}
\alias{hyper_index}
\alias{hyper_index}
\alias{hyper_index.tbl_df}
\alias{hyper_index}
\alias{hyper_index.tidync}
\alias{hyper_index}
\alias{hyper_index.character}
\alias{hyper_index}
\alias{hyper_index.hyperfilter}
\title{hyper slab index}
\usage{
hyper_index(x, ...)

\method{hyper_index}{tbl_df}(x, ...)

\method{hyper_index}{tidync}(x, ...)

\method{hyper_index}{character}(x, varname, ...)

\method{hyper_index}{hyperfilter}(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{expressions to \code{hyper_filter}}

\item{varname}{variable to be activated}
}
\description{
hyper slab index
}
