% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{hyper_slice}
\alias{hyper_slice}
\alias{hyper_slice}
\alias{hyper_slice.hyperindex}
\alias{hyper_slice}
\alias{hyper_slice.hyperfilter}
\alias{hyper_slice}
\alias{hyper_slice.NetCDF}
\alias{hyper_slice}
\alias{hyper_slice.character}
\title{hyper slice}
\usage{
hyper_slice(x, ...)

\method{hyper_slice}{hyperindex}(x, ...)

\method{hyper_slice}{hyperfilter}(x, ...)

\method{hyper_slice}{NetCDF}(x, ...)

\method{hyper_slice}{character}(x, ...)
}
\arguments{
\item{x}{a hyperindex-able thing}

\item{...}{ignored}
}
\description{
hyper slice
}
