% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_slice.R
\name{hyper_slice}
\alias{hyper_slice}
\alias{hyper_slice}
\alias{hyper_slice.tidync}
\alias{hyper_slice}
\alias{hyper_slice.character}
\title{hyper slice}
\usage{
hyper_slice(x, select_var = NULL, ..., raw_datavals = FALSE,
  force = FALSE)

\method{hyper_slice}{tidync}(x, select_var = NULL, ...,
  raw_datavals = FALSE)

\method{hyper_slice}{character}(x, ..., select_var = NULL,
  raw_datavals = FALSE)
}
\arguments{
\item{x}{tidync object}

\item{select_var}{optional vector of variable names to select}

\item{...}{ignored}

\item{raw_datavals}{logical to control whether scaling in the NetCDF is applied or not}

\item{force}{ignore caveats about large extraction and just do it}
}
\description{
By default all variables in the active grid are returned, use \code{select_var} to limit.
}
