% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-tbl_cube.R
\name{hyper_tbl_cube}
\alias{hyper_tbl_cube}
\alias{hyper_tbl_cube}
\alias{hyper_tbl_cube.tidync}
\alias{hyper_tbl_cube}
\alias{hyper_tbl_cube.character}
\alias{hyper_tbl_cube}
\alias{hyper_tbl_cube.hyperfilter}
\title{dplyr tbl cube from NetCDF}
\usage{
hyper_tbl_cube(x, ...)

\method{hyper_tbl_cube}{tidync}(x, ...)

\method{hyper_tbl_cube}{character}(x, ...)

\method{hyper_tbl_cube}{hyperfilter}(x, ...)
}
\arguments{
\item{x}{filename}

\item{...}{arguments for \code{hyper_filter}}
}
\description{
dplyr tbl cube from NetCDF
}
