% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-tbl_cube.R
\name{hyper_tbl_cube}
\alias{hyper_tbl_cube}
\alias{hyper_tbl_cube}
\alias{hyper_tbl_cube.tidync}
\alias{hyper_tbl_cube}
\alias{hyper_tbl_cube.character}
\title{dplyr tbl cube}
\usage{
hyper_tbl_cube(x, ...)

\method{hyper_tbl_cube}{tidync}(x, ...)

\method{hyper_tbl_cube}{character}(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{arguments for \code{hyper_filter}}
}
\value{
tbl_cube

\code{dplyr::tbl_cube}
}
\description{
Produce a dplyr \code{tbl_cube} from NetCDF.
}
\details{
#'
}
\examples{
ufile <- system.file("extdata", "unidata", "test_hgroups.nc", package = "tidync")
tidync(ufile)
hyper_tbl_cube(tidync(ufile))
f <- "S20092742009304.L3m_MO_CHL_chlor_a_9km.nc"
l3file <- system.file("extdata/oceandata", f, package= "tidync")
(cube <- hyper_tbl_cube(tidync(l3file) \%>\%
activate(chlor_a), lon = lon > 107, lat = abs(lat) < 30))
# library(ggplot2)
# scl <- function(x) {rng <- range(x, na.rm = TRUE); (x - rng[1])/diff(rng)}
# library(palr)
# pal <- chlPal(palette = TRUE)
# sfg <- scale_fill_gradientn(values = scl(head(pal$breaks, -1)),
#      colours = pal$cols)
# ggplot(tibble::as_tibble(cube) \%>\% dplyr::filter(!is.na(chlor_a))) +
#geom_raster(aes(lon, lat, fill = chlor_a)) + sfg  + coord_equal() #+
#geom_polygon(data = ggplot2::map_data("world",
#xlim = c(100, 180),
#ylim = c(-30, 30)), aes(long, lat, group = group, fill = NULL))
}
