% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{hyper_tibble}
\alias{hyper_tibble}
\alias{hyper_tibble}
\alias{hyper_tibble.character}
\alias{hyper_tibble}
\alias{hyper_tibble.NetCDF}
\alias{hyper_tibble}
\alias{hyper_tibble.hyperfilter}
\title{hyper tibble}
\usage{
hyper_tibble(x, ...)

\method{hyper_tibble}{character}(x, ...)

\method{hyper_tibble}{NetCDF}(x, ...)

\method{hyper_tibble}{hyperfilter}(x, ...)
}
\arguments{
\item{x}{object to tibbulate}

\item{...}{arguments to `hyper_filter``}
}
\value{
a `tbl_df`
}
\description{
hyper tibble
}
\examples{
f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncdump")
hyper_filter(f)
library(dplyr)
hyper_slice(f, lat = lat > 0) \%>\% dim()

  \%>\% filter(!is.na(chlor_a)) 
ht   
library(ggplot2)
ggplot(ht \%>\% filter(!is.na(chlor_a)), 
aes(x = lon, y = lat, fill = chlor_a)) + geom_pch()

}
