% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_tibble.R
\name{hyper_tibble}
\alias{hyper_tibble}
\alias{hyper_tibble}
\alias{hyper_tibble.character}
\alias{hyper_tibble}
\alias{hyper_tibble.tidync}
\alias{hyper_tibble}
\alias{hyper_tibble.hyperfilter}
\title{hyper tibble}
\usage{
hyper_tibble(x, ...)

\method{hyper_tibble}{character}(x, ...)

\method{hyper_tibble}{tidync}(x, ...)

\method{hyper_tibble}{hyperfilter}(x, ...)
}
\arguments{
\item{x}{object to tibbulate}

\item{...}{arguments to `hyper_filter``}
}
\value{
a \code{tbl_df}
}
\description{
hyper tibble
}
\examples{
l3file <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package= "ncdump")
rnc <- tidync(l3file)
hyper_filter(rnc)
library(dplyr)
hyper_slice(l3file, lat = lat > 0) \%>\% dim()

 ht <- hyper_tibble(rnc) \%>\% filter(!is.na(chlor_a)) 
ht   
library(ggplot2)
ggplot(ht \%>\% filter(!is.na(chlor_a)), 
aes(x = lon, y = lat, fill = chlor_a)) + geom_point()

}
