% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_tibble.R
\name{hyper_tibble}
\alias{hyper_tibble}
\alias{hyper_tibble}
\alias{hyper_tibble.character}
\alias{hyper_tibble}
\alias{hyper_tibble.tidync}
\title{hyper tibble}
\usage{
hyper_tibble(x, ..., na.rm = TRUE)

\method{hyper_tibble}{character}(x, ..., na.rm = TRUE)

\method{hyper_tibble}{tidync}(x, ..., na.rm = TRUE)
}
\arguments{
\item{x}{object to tibbulate}

\item{...}{arguments to `hyper_filter``}

\item{na.rm}{if \code{TRUE} these rows are not included in the output when all variables are \code{NA}}
}
\value{
a \code{tbl_df}
}
\description{
hyper tibble
}
\examples{
l3file <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
f <- system.file("extdata", "oceandata", l3file, package= "tidync")
rnc <- tidync(f)
hyper_filter(rnc)
library(dplyr)
lapply(hyper_slice(f, lat = lat > 0, lon = index > 3000), dim)

 ht <- hyper_tibble(rnc) \%>\% 
 filter(!is.na(chlor_a)) 
ht   
library(ggplot2)
ggplot(ht \%>\% filter(!is.na(chlor_a)), 
aes(x = lon, y = lat, fill = chlor_a)) + geom_point()

}
