% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_utils.R
\name{hyper_vars}
\alias{hyper_vars}
\alias{hyper_vars}
\alias{hyper_dims}
\title{Hyper utilities}
\usage{
hyper_vars(x, ...)

hyper_dims(x, ...)
}
\arguments{
\item{x}{tidync object}

\item{...}{ignored}
}
\value{
data frame
}
\description{
Functions to report on the current \code{hyper_filter} status of the \code{active}
grid.
}
\details{
\code{hyper_vars} will list the names of the variables on the currently active grid
\code{hyper_dims} will list the ids, names, lengths, and start and count records
on the currently active dimensions
}
\examples{
f <- "S20080012008031.L3m_MO_CHL_chlor_a_9km.nc"
l3file <- system.file("extdata/oceandata", f, package= "tidync")
tnc <- tidync(l3file)
hyper_vars(tnc)
hyper_dims(tnc)
hyper_dims(tnc \%>\% hyper_filter(lat = lat < 20))
}
