% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf-utils.R
\name{nc_get}
\alias{nc_get}
\title{Helper to get a variable from NetCDF.}
\usage{
nc_get(x, v, test = FALSE)
}
\arguments{
\item{x}{file path, uri, or NetCDF connection}

\item{v}{variable name}

\item{test}{if true we make sure the connection can be open, not applied for connections themselves}
}
\description{
This exists so we can (internally) use a file path, uri, or open
NetCDF connection (ncdf4 or RNetCDF) in a simpler way.
}
\details{
This function just reads the whole array. It is equivalent to the
angstroms package function 'rawdata(x, varname)'.
}
