% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{print.tidync}
\alias{print.tidync}
\title{Print tidync object}
\usage{
\method{print}{tidync}(x, ...)
}
\arguments{
\item{x}{NetCDF object}

\item{...}{reserved}
}
\description{
Provide a summary of variables and dimensions, organized by their 'grid' (or
'shape') and with a summary of any slicing operations provided as 'start' and
'count' summaries for each dimension in the active grid.
}
\details{
The print summary is organized in two sections, the first is available grids
(sets of dimensions) and their associated variables, the second is the
dimensions, separated into active and inactive. All dimensions may be active
in some NetCDF sources.

Individual \emph{active} dimensions include the following components:
\itemize{
\item 'dim' - dimension label, D0, D1, D2, ...
}
}
