% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{print.tidync}
\alias{print.tidync}
\title{Print tidync object}
\usage{
\method{print}{tidync}(x, ...)
}
\arguments{
\item{x}{NetCDF object}

\item{...}{reserved}
}
\description{
Provide a summary of variables and dimensions, organized by their 'grid' (or
'shape') and with a summary of any slicing operations provided as 'start' and
'count' summaries for each dimension in the active grid.
}
