% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_tidyfile.R
\name{tidyfile}
\alias{tidyfile}
\alias{tidyfile}
\alias{tidyfile.data.frame}
\alias{tidyfile}
\alias{tidyfile.raadfiles}
\alias{tidyfile}
\alias{tidyfile.character}
\title{Tidy file collection}
\usage{
tidyfile(x, all = FALSE, ...)

\method{tidyfile}{data.frame}(x, all = FALSE, ...)

\method{tidyfile}{raadfiles}(x, all = FALSE, ...)

\method{tidyfile}{character}(x, all = FALSE, ...)
}
\arguments{
\item{x}{raadfiles, data frame with "fullname", or character vector}

\item{all}{scan metadata from all files?}

\item{...}{ignored}
}
\value{
data frame of filename "fullname" and "meta" in a list-col
}
\description{
Tidy files include a trick to only store the full metadata of the
first file in the collection. This is extracted using \code{ncmeta::nc_meta} in
a form that may changed substantially. The current plan is to always
have an \emph{estimable} set of metadata using the first file as a sample,
but to be able to scan the full metadata as required.
This is based on the model where a set of files \emph{do match} or can be expected
to match in every shape other than the \emph{unlimited} dimension, the one
that is dispersed over files (usually "time").
}
\examples{
#library(dplyr)
#files <- raadfiles::oisst_daily_files() \%>\% dplyr::filter(as.Date(date) > (Sys.Date() - 40))
#f <- tidyfile(files, all = FALSE)
}
