% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync-package.r, R/tidync.R
\docType{package}
\name{tidync}
\alias{tidync}
\alias{tidync-package}
\alias{tidync}
\title{Tidy tools for NetCDF data.}
\usage{
tidync(x, ...)
}
\arguments{
\item{x}{path to a NetCDF file}
}
\description{
Provides easy to use idioms for working with NetCDF data for
extraction, manipulation and visualization. 
NetCDF is Network Common Data Form \url{https://www.unidata.ucar.edu/software/netcdf/}.

Function to extract all metadata from a NetCDF, for use in subsequent operations.
}
\details{
There is a family of functions "hyper_verb" around exploring and
extracting data.  
\tabular{ll}{
 \code{\link{hyper}} \tab extracts all file metadata in an object useable by other tidync functions\cr
 \code{\link{hyper_filter}} \tab processes user-driven expressions to filter based on NetCDF dimensions \cr
 \code{\link{hyper_index}} \tab processes dimension filters into NetCDF extraction indexes \cr
 \code{\link{hyper_slice}} \tab extracts a raw data array based on a NetCDF index \cr
 \code{\link{hyper_tibble}} \tab extracts the slice as a data frame with all dimension values \cr
 }

Any NetCDF with variable arrays should work. Files with compound types are not yet supported. We
haven't explored HDF5 per se, so any feedback is appreciated.
}
