% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{tidync}
\alias{tidync}
\alias{tidync}
\alias{tidync.character}
\title{tidy netcdf}
\usage{
tidync(x, what, ...)

\method{tidync}{character}(x, what, ...)
}
\arguments{
\item{x}{path to a NetCDF file}

\item{what}{(optional) character name of grid (see \code{ncmeta::nc_grids}) or (bare) name of variable (see \code{ncmeta::nc_vars}) or index of grid to \code{activate}}

\item{...}{reserved for arguments to methods, currently ignored}
}
\description{
Function to extract all metadata from a NetCDF, for use in subsequent operations. By default
the first \emph{shape} encountered is  \code{activate}d.
}
\details{
Any NetCDF with variable arrays should work. Files with compound types are not yet supported and
should fail gracefully.

We haven't yet explored HDF5 per se, so any feedback is appreciated. Major
use of compound types is made by https://github.com/mdsumner/roc
}
\examples{
#cf <- raadtools::cpolarfiles()
#nc <- tidync(cf$fullname[1])
#print(nc)
}
