% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{tidync}
\alias{tidync}
\title{tidy netcdf}
\usage{
tidync(x, ...)
}
\arguments{
\item{x}{path to a NetCDF file}

\item{...}{unused argments}
}
\description{
Function to extract all metadata from a NetCDF, for use in subsequent operations.
}
\details{
Any NetCDF with variable arrays should work. Files with compound types are not yet supported. We
haven't explored HDF5 per se, so any feedback is appreciated.
}
