% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{tidync}
\alias{tidync}
\alias{tidync}
\alias{tidync.character}
\alias{tidync}
\alias{print.tidync}
\title{tidy netcdf}
\usage{
tidync(x, ...)

\method{tidync}{character}(x, what)

\method{print}{tidync}(x, ...)
}
\arguments{
\item{x}{path to a NetCDF file}

\item{...}{reserved}

\item{what}{(optional) character or bare name of variable to \code{activate}}

\item{x}{NetCDF object}
}
\description{
Function to extract all metadata from a NetCDF, for use in subsequent operations. By default
the first \emph{shape} encountered is  \code{activate}d.

print S3 method
}
\details{
Any NetCDF with variable arrays should work. Files with compound types are not yet supported. We
haven't explored HDF5 per se, so any feedback is appreciated.

Prints a summary of variables and dimensions, organized by their 'shape' - "same shape" means "same grid".
Shape is an instance of a particular set of dimensions, which can be shared by more than one
variable. This is not the 'rank' of a variable (the number of dimensions) since a single
data set may have many 3D variables composed of different sets of axes/dimensions. There's no
formality around the concept of 'shape', as far as we know.

A dimension may have length zero, but this is a special case for a "measure" dimension, we think. (It
doesn't mean the product of the dimensions is zero, for example).
}
\examples{
cf <- raadtools::cpolarfiles()
nc <- tidync(cf$fullname[1])
print(nc)
}
