% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidync.R
\name{tidync}
\alias{tidync}
\title{tidy netcdf}
\usage{
tidync(x, what)
}
\arguments{
\item{x}{path to a NetCDF file}

\item{what}{(optional) character or bare name of variable to \code{activate}}
}
\description{
Function to extract all metadata from a NetCDF, for use in subsequent operations. By default
the first variable encountered is  \code{activate}d.
}
\details{
Any NetCDF with variable arrays should work. Files with compound types are not yet supported. We
haven't explored HDF5 per se, so any feedback is appreciated.
}
